Refactor flow controller for reuse in opening workspace
